+++ lib/src/descriptor.pbserver.dart
@@ -1,0 +1,7 @@
+///
+//  Generated code. Do not modify.
+///
+library proto2_descriptor_pbserver;
+
+export 'descriptor.pb.dart';
+
