+++ lib/src/plugin.pbenum.dart
@@ -1,0 +1,5 @@
+///
+//  Generated code. Do not modify.
+///
+library proto2.compiler_plugin_pbenum;
+
