+++ lib/src/plugin.pbserver.dart
@@ -1,0 +1,7 @@
+///
+//  Generated code. Do not modify.
+///
+library proto2.compiler_plugin_pbserver;
+
+export 'plugin.pb.dart';
+
